Putchar now always writes to console, never to fd 1.
This introduces a semantic difference between fprintf(1, "xxx");
and printf("xxx"); when fd 1 is not the console, but the
difference is tremendously convenient when debugging
the file descriptors themselves.

This `new' behavior is the same behavior we had last
year.  Recent reorgs to the printf code made it go away
temporarily.