all the jos code uses
type
nameoffunction(args)
{
	...
}

let's keep doing that.
it looks sloppy when we mix styles indiscriminately.

also, all the comments should be // C++ style
except in headers that get included by assembly.
i'm not defending things, just saying that that's
the way they are and it looks sloppy to start mixing.