More checks during page_check, some inspired by Don Porter at U. Texas.
Perhaps the most significant is that page_check now
starts off setting the first 128 bytes of every free
page to 0x97.  This will catch students who don't
clear the page in pgdir_walk very early.  (An equally
good alternative is to clear the page in page_alloc.)
This in turn should avoid some of the more mysterious
bugs in future labs.

Perhaps we should suggest in the comments that
page_alloc zero the page?