Discard the .comment section.  This fixes tons of failing tests.
Why?  Without this, the linker places the .comment section before the
.stabs section.  Because the .comment input section is never
explicitly mentioned in the linker script, the linker places it at VMA
0, which resets . to 0, which places the .stabs section at VMA 0 plus
some change.  The resulting ELF file has a loadable segment with
virtual address 0 that contains .comment, .stabs, and .stabstr.  The
kernel maps this segment in, allocating memory at virtual address 0.
As a result, you *can* write to VA 0, which causes all of the user
tests that are supposed to produce page faults by reading or writing
to VA 0 to produce no fault, and thus fail.